docs: Add CLI reference documentation
This adds a hidden command to generate markdown documentation for the
CLI usage. This could be used to pull in to a rendered web site, but for
now it provides nicely rendered files with navigable hyperlinks when
viewing via GitHub's rendered markdown views.

It also adds a GitHub Action (disclaimer: not fully tested so may need
some tweaks) that checks on any merged updates to the CLI on the main
branch. If it detects that there are CLI documentation changes it will
create a pull request to update the markdown files.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>